Рефакторинг модуля с  утилитами
