+++ src/mock/point.js
@@ -1,4 +1,4 @@
-import {getRandomInteger} from "../utils.js";
+import {getRandomInteger} from "../utils/common.js";
 
 // Генерирует случайный тип маршрута
 const generateType = () => {
